+++ tests/testthat/test_manage.R
@@ -1,5 +1,3 @@
-context("Data management")
-
 skip_on_cran()